WebSockets/tests/submissions/Opera/cookies/002.html

<!doctype html>
<title>WebSockets: Set-Cookie in response</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/set-cookie');
  ws.onopen = t.step_func(function(e) {
    assert_regexp(document.cookie, /ws_test=test/);
    // remove cookie
    document.cookie = 'ws_test=; Path=/; Expires=Sun, 06 Nov 1994 08:49:37 GMT';
    ws.close();
    t.done();
  });
  ws.onerror = ws.onclose = t.step_func(function(e) {assert_unreached(e.type)});
});
</script>
